cmd: utils.py for watcher
