+++ nb_workflows/cmd/executors.py
@@ -10,7 +10,8 @@ from nb_workflows.client.diskclient import DiskClient
 from nb_workflows.conf import defaults, load_client
 from nb_workflows.executors.context import pure_execid
 from nb_workflows.types import NBTask
-from nb_workflows.utils import format_bytes
+
+from .utils import watcher
 
 console = Console()
 
@@ -21,30 +22,6 @@ def _parse_params_args(params):
         k, v = p.split("=")
         params_dict.update({k: v})
     return params_dict
-
-
-def watcher(c: DiskClient, execid, stats=False):
-    keep = True
-    last = None
-
-    console.print(f"[bold magenta]Watching for execid: {execid}[/]")
-    events = 0
-
-    for evt in c.events_listen(execid, last=last):
-        if evt.event == "stats" and stats:
-            _stats = json.loads(evt.data)
-            _mem = _stats["mem"]["mem_usage"]
-            mem = format_bytes(_mem)
-            msg = f"[orange]Memory used {mem}[/]"
-            console.print(msg)
-        elif evt.event == "result":
-            keep = False
-            console.print(f"[bold green]Finished execid: [/] [bold magenta]{execid}[/]")
-        elif evt.event != "stats":
-            msg = evt.data
-            console.print(msg)
-        last = evt.id
-        events += 1
 
 
 @click.group(name="exec")
